fargate/requirements.txt
aws-cdk-lib==2.78.0
constructs==10.2.15
